tempsense: Fix PDN bug that caused LVS to fail (workaround)
Pdngen didn't assign the VPWR and VPB pins in the TEMP_ANALOG
voltage domain correctly to VIN (instead of VDD). Although
the resulting layout was correct, this produced a mismatch
between the netlist generated by OpenROAD and the one extracted
from the GDS file with Magic.

This commit adds a block of code to global_place.tcl that manually
assigns these pins for the instances in the TEMP_ANALOG domain.

When pdngen gets updated to the new .tcl format (issue #81), this
block of code should be removed since it won't be necessary.