tempsense: Move customPlace proc to separate file, fix cell orient
Moved the customPlace_east proc (#77) to a separate file for better
organization (scripts/openfasoc/custom_place.tcl).

Also fixes the bug where this proc would put the cells in their
respective rows, sequentially, but not flip them in subsequent
rows (need to be flipped because VDD and GND flip from one row
to the next).

Finally, also changes the status of placed cells to FIRM to ensure
they won't move when detailed_placement is called.