tempsense: add 'isfixed' option to customPlace_east command
Allows the user to choose if the cells should be placed exactly in
the rows defined, or if they can be moved slightly during detailed
placement, to avoid overlapping other fixed cells (for ex., taps).

Using isfixed as "no", this procedure places the cells in a region
around the east of the design, while "yes" puts them exactly at
the start of the Nth row (as given) subsequently.