tempsense: update config.mk, git ignore simulations/run/
This changes the assignment of some of the varaibles in the design
config.mk file. It makes them export with '=?' instead of '=', to
allow the variables to be set >>before<< the generator is run,
from outside OpenFASOC, without having to change the generator files.

Also adds simulations/run/ directory to .gitignore