gdsfactory-gen: Added fix for overlaps in the straight route via placements
Via placements in the straight route code led to a few buggy overlaps in layouts.
Orientation of the vias is now taken as input from the programmer or is according
to the orientation of the port by default. Added relative import fixes for op-amp
and straight route standalone testing as well. Tested locally on op-amp gen.

Signed-off-by: L Lakshmanan <karthiklaksh1729@gmail.com>